fixes and a simple test program (prettylisten) for mig
git-svn-id: http://tinyos-main.googlecode.com/svn@1895 285f258d-8064-b571-fbc6-9656dc998e90

+++ branches/tinyos-2_0_devel-BRANCH/support/sdk/c/message.h
@@ -19,10 +19,10 @@ typedef struct tmsg tmsg_t;
 
 void tmsg_fail(void);
 
-tmsg_t *new_tmsg(void *packet, int len);
+tmsg_t *new_tmsg(void *packet, size_t len);
 void free_tmsg(tmsg_t *msg);
 void *tmsg_data(tmsg_t *msg);
-int tmsg_length(tmsg_t *msg);
+size_t tmsg_length(tmsg_t *msg);
 
 uint64_t tmsg_read_ule(tmsg_t *msg, size_t bit_offset, size_t bit_length);
 int64_t tmsg_read_le(tmsg_t *msg, size_t bit_offset, size_t bit_length);
